[Python] Fixed import when AvroSchema is not being used
